+++ test/Elastica/Index/SettingsTest.php
@@ -59,7 +59,7 @@ class SettingsTest extends BaseTest
     /**
      * @group functional
      */
-    public function testSetNumberOfReplicas()
+    public function testSetGetNumberOfReplicas()
     {
         $indexName = 'test';
 
@@ -68,13 +68,37 @@ class SettingsTest extends BaseTest
         $index->create([], true);
         $settings = $index->getSettings();
 
-        $settings->setNumberOfReplicas(2);
+        // Check for zero replicas
+        $settings->setNumberOfReplicas(0);
         $index->refresh();
-        $this->assertEquals(2, $settings->get('number_of_replicas'));
+        $this->assertEquals(0, $settings->get('number_of_replicas'));
+        $this->assertEquals(0, $settings->getNumberOfReplicas());
 
+        // Check with 3 replicas
         $settings->setNumberOfReplicas(3);
         $index->refresh();
         $this->assertEquals(3, $settings->get('number_of_replicas'));
+        $this->assertEquals(3, $settings->getNumberOfReplicas());
+
+        $index->delete();
+    }
+
+    /**
+     * @group functional
+     */
+    public function testGetNumberOfReplicas()
+    {
+        $indexName = 'test';
+
+        $client = $this->_getClient();
+        $index = $client->getIndex($indexName);
+        $index->create([], true);
+
+        $settings = $index->getSettings();
+
+        // Test with default number of replicas
+        $this->assertEquals(IndexSettings::DEFAULT_NUMBER_OF_REPLICAS, $settings->get('number_of_replicas'));
+        $this->assertEquals(IndexSettings::DEFAULT_NUMBER_OF_REPLICAS, $settings->getNumberOfReplicas());
 
         $index->delete();
     }